requirements.txt
django==2.2
pillow
djangorestframework-simplejwt==4.5.0
djangorestframework
markdown
django-filter
django-crispy-forms
django-reversion
django-formtools
django-import-export==2.0
future
httplib2
six
requests
coreapi
django-cors-headers
xlwt
xlsxwriter
drf-extensions
django-redis
pycryptodome
raven
social-auth-core==3.4.0
social-auth-app-django==3.4.0